% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.R
\name{algo}
\alias{algo}
\title{Run the algorithm}
\usage{
algo(pars, which_alpha_pars, alpha_pars, alpha, g, xs, p_log,
  which_pars_perturbed)
}
\arguments{
\item{pars}{pars}

\item{which_alpha_pars}{names of the alpha pars used}

\item{alpha_pars}{alpha pars}

\item{alpha}{numeric 1L.}

\item{g}{observation funciton}

\item{xs}{steady state prediction function}

\item{p_log}{parameter trafo}

\item{which_pars_perturbed}{character. which pars are perturbed}
}
\value{
a tibble with r0, r1, rkep and ropt
}
\description{
This function has a lot of arguments, to ensure that each object is the correct one
}
