% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{apply_expression}
\alias{apply_expression}
\title{magrittr-Style Apply an expression that's more complicated}
\usage{
apply_expression(x, ...)
}
\arguments{
\item{expr}{}
}
\description{
magrittr-Style Apply an expression that's more complicated
}
