% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_preturbations.R
\name{cluster_matrices_and_print_perturbations}
\alias{cluster_matrices_and_print_perturbations}
\title{clustering and print function}
\usage{
cluster_matrices_and_print_perturbations(results_unnested_matrices, a,
  col_start, nclust, which_clust)
}
\arguments{
\item{results_unnested_matrices}{result from above}

\item{a}{alpha, the value at which the local response matrices are compared}

\item{col_start}{regex like "up_" or "both_"}

\item{nclust}{nclusters}

\item{which_clust}{print perturbations of which cluster?}
}
\description{
clustering and print function
}
