% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{combine_dr_list}
\alias{combine_dr_list}
\title{Combine a list of dose_response data.frames with different pars_opt combinations}
\usage{
combine_dr_list(dr_list)
}
\arguments{
\item{dr_list}{}
}
\description{
Combine a list of dose_response data.frames with different pars_opt combinations
}
