% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{dose_response_fun}
\alias{dose_response_fun}
\title{Compute dose-response curves}
\usage{
dose_response_fun(which_par, dosages, pars = pars_0, predfun = (xs * p_log))
}
\arguments{
\item{which_par}{Character of length 1: Which par to perturb}

\item{dosages}{Numeric including the dosages eq(1,50,by = 1)}

\item{predfun}{}
}
\description{
Compute the steady state dose-response of a parameter.
Put the dose response out in a tidy data format.
}
