% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{find_symmetric_elements_both_nonzero}
\alias{find_symmetric_elements_both_nonzero}
\title{Find elements Aij and Aji which are both non-zero}
\usage{
find_symmetric_elements_both_nonzero(mat, threshold = 0.01)
}
\arguments{
\item{mat}{a square matrix}
}
\description{
Find elements Aij and Aji which are both non-zero
}
