% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_preturbations.R
\name{get_r0_r1_ropt_rkept}
\alias{get_r0_r1_ropt_rkept}
\title{Get the relevant matrices for a perturbation and alpha-setting}
\usage{
get_r0_r1_ropt_rkept(results, a, pp, pars, xs, obs_fun, p_log)
}
\arguments{
\item{results}{as is from runbg}

\item{a}{alpha value}

\item{pp}{names of perturbed parameters}

\item{pars, xs, obs_fun, p_log}{objects that were used to generate the result}
}
\value{
list of matrices
}
\description{
Get the relevant matrices for a perturbation and alpha-setting
}
