% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{global_response_matrix}
\alias{global_response_matrix}
\title{Global response matrix R as a function of the perturbed parameters}
\usage{
global_response_matrix(pars_perturbed, pars)
}
\arguments{
\item{pars}{}
}
\description{
This version calculates d ln(x_i) /dp_j with finite differences as derivatives
}

