% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{global_response_matrix_fractional}
\alias{global_response_matrix_fractional}
\title{Global response matrix R}
\usage{
global_response_matrix_fractional(pars_perturbed, pars)
}
\arguments{
\item{pars}{}
}
\description{
This version calculates d ln x_i/dp_j as suggested in the paper
}

