% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonedfromdModandcOde.R
\name{mname.function}
\alias{mname.function}
\title{This function is needed such that the new mstrust works with functions other than fn's}
\usage{
\method{mname}{function}(x, conditions = NULL)
}
\description{
This function is needed such that the new mstrust works with functions other than fn's
}
