% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{obj_alpha}
\alias{obj_alpha}
\title{Objective function to minimize certain elements of the local response matrix}
\usage{
obj_alpha(pars_opt = pars_opt_0,
  perturbation_prediction = perturbation_prediction_0, r_kept = r_kept_0,
  obs_fun = g, p_fun = p_pert, mypars = pars_0, fixed = NULL)
}
\arguments{
\item{pars}{}
}
\value{
The parameter values for which the sign-changing elements are minimal.
}
\description{
This function squares all elements that flip sign once the complexes are added to the observation.
This can then be used to minimize the values.
}

