% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{obj_frobenius}
\alias{obj_frobenius}
\title{Objective function to minimize certain elements of the local response matrix}
\usage{
obj_frobenius(pars_optimization, obj_pars, r_sign_changed)
}
\arguments{
\item{pars_optimization}{The parameters which are needed to optimize the matrix}

\item{r_sign_changed}{A matrix of logicals specifying the elements to minimize.}
}
\value{
The parameter values for which the sign-changing elements are minimal.
}
\description{
This function squares all elements that flip sign once the complexes are added to the observation.
This can then be used to minimize the values.
}

