% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_generation.R
\name{p_ic_fun}
\alias{p_ic_fun}
\title{Automatically generaty parameter trafos to quickly test different parameter values as conditions}
\usage{
p_ic_fun(ic_list, pars = pars_0, mixed = FALSE)
}
\arguments{
\item{ic_list}{Named list of parameter values to be set.}

\item{pars}{The original set of parameters to be replaced.}

\item{mixed}{Should all combinations of the pars be gone through (TRUE)
or should the other pars be set to their first value supplied, when scanning through a certain parameter?}
}
\value{
A "parfn" of all conditions.
}
\description{
This function can be used, if one wants quickly wants to check the behaviour of the model with different parameters.
The named list ic_list contains all the parameter values that one wants to try out.
The function then returns a set of parameter transformations p_ic which can be put into g*x*p_ic to have a look at the model prediction.
}
