% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_generation.R
\name{P_log}
\alias{P_log}
\title{Generate a log-trafo}
\usage{
P_log(parameters = NULL, condition = NULL, attach.input = FALSE,
  keep.root = TRUE, compile = FALSE, modelname = NULL,
  method = c("explicit", "implicit"), verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Just a convenience-function to quickly generate a log-trafo.
}
