% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_generation.R
\name{p_pert_fun}
\alias{p_pert_fun}
\title{Convenient simulation of perturbation experiments}
\usage{
p_pert_fun(pars_perturbed, pars, modelname = NULL)
}
\arguments{
\item{pars_perturbed}{A named vector with the changes on the log scale of the perturbed parameters.}

\item{pars}{All parameters of the model.}
}
\value{
A parameter trafo function of class parfn, which goes from pars_outer (log scale) to pars_outer (log scale)
}
\description{
A parameter transformation function to conveniently generate conditions, in which on parameter was perturbed
}
\details{
Attention: if !is.null(modelname) you have to run compile(p_pert) after calling p_pert <- p_pert_fun(..)
}
