% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_response_matrix.R
\name{print.r_small}
\alias{print.r_small}
\title{Print the local response matrix without R and steady_state attributes.}
\usage{
\method{print}{r_small}(x)
}
\arguments{
\item{x}{}
}
\description{
Print the local response matrix without R and steady_state attributes.
}

