% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{r_alpha_fun}
\alias{r_alpha_fun}
\title{Just a convenience function for r(alpha)}
\usage{
r_alpha_fun(pars_opt = pars_opt_0,
  perturbation_prediction = perturbation_prediction_0, obs_fun = g0,
  p_fun = (p_log * p_pert), pars = pars_0)
}
\arguments{
\item{pars}{}
}
\description{
Just a convenience function for r(alpha)
}
