% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_response_matrix.R
\name{r_elements_kept_fun}
\alias{r_elements_kept_fun}
\title{Determine which signs have flipped over by including the complexes}
\usage{
r_elements_kept_fun(pars = pars_0, predfun = (g * xs),
  which_pars_perturbed = which_pars_perturbed_0,
  pars_optimization = pars_opt_0, r_fun = r_numDeriv_fun)
}
\arguments{
\item{which_pars_perturbed}{}
}
\description{
To do: have a look at what happens, if pars_opt = .5 or combinations of pars_opt (eg c(0,1), c(1,0) ...)
}

