% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{r_kept_fun}
\alias{r_kept_fun}
\title{Which elements to keep in the optimization procedure?}
\usage{
r_kept_fun(pars_opt = pars_opt_0,
  perturbation_prediction = perturbation_prediction_0, obs_fun = g0,
  p_fun = (p_log * p_pert), pars = pars_0, alpha = 9)
}
\arguments{
\item{pars}{}
}
\description{
This function compares the connection coefficient in the local response matrices
when either active forms or complexes are used as a communicating species.
It takes two observation functions, one with complex being the "standard" cs and one with the free form bein the standard cs.
}
\details{
"Standard" in the sense from above: m2+a*cme -> m2 is the standard in this one.
}
