% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{r_kept_fun2}
\alias{r_kept_fun2}
\title{r_kept fun for supplying two r-matrices}
\usage{
r_kept_fun2(r_0, r_alpha)
}
\arguments{
\item{r_alpha}{}
}
\description{
r_kept fun for supplying two r-matrices
}
