% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_response_matrix.R
\name{r_numDeriv_fun}
\alias{r_numDeriv_fun}
\title{Title}
\usage{
r_numDeriv_fun(pars = pars_0, predfun = (g * xs),
  which_pars_perturbed = which_pars_perturbed_0, pars_optimization = NULL)
}
\arguments{
\item{pars_optimization}{}
}
\description{
This function only works for one condition
}

