% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{r_opt_dose_response}
\alias{r_opt_dose_response}
\title{Compute dose responses for one parameter}
\usage{
r_opt_dose_response(which_par, dosages, alpha_scan_range = seq(-9, 9, by = 2),
  pars_opt = pars_opt_0, obs_fun = g, p_fun = (p_log * p_pert),
  pars = pars_0)
}
\arguments{
\item{pars}{}
}
\description{
Optimize whenever a sign flip occurs
}
