% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{r_opt_dose_response_separate_scanning}
\alias{r_opt_dose_response_separate_scanning}
\title{Compute dose responses for one parameter, scan pars_opt separately}
\usage{
r_opt_dose_response_separate_scanning(which_par, dosages,
  alpha_scan_range = seq(-9, 9, by = 2), which_par_opt, pars_opt,
  obs_fun, p_fun, pars)
}
\arguments{
\item{which_par}{Which dynamical parameter shall be scanned?}

\item{dosages}{Scan range of this parameter}

\item{alpha_scan_range}{Which values of free parameters a shall be scanned?}

\item{pars_opt}{Free parameters a_ij to control the influence of the complexes}

\item{obs_fun}{Observation function which returns the communicating species}

\item{p_fun}{Parameter transformation which returns different conditions}

\item{pars}{Parameters of the ODE model}
}
\description{
Scan each free parameter a_ij individually while keeping others at fixed values.
Optimize whenever a sign flip occurs
}
