% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{r_opt_fun}
\alias{r_opt_fun}
\title{Local response matrix r as a function of the some explicitly specified parameters}
\usage{
r_opt_fun(pars_optimization, pars)
}
\arguments{
\item{pars_optimization}{}
}
\description{
The local response matrix r is computed with the standard set of parameters "pars".
If a parameter is given in pars_optimization, this value is taken instead.
}

