% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{r_sign_changed_fun}
\alias{r_sign_changed_fun}
\title{Determine which signs have flipped over by including the complexes}
\usage{
r_sign_changed_fun(pars_optimization, pars)
}
\arguments{
\item{pars}{}
}
\description{
To do: have a look at what happens, if pars_opt = .5 or combinations of pars_opt (eg c(0,1), c(1,0) ...)
}

