% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_generation.R
\name{reduce_ic_list}
\alias{reduce_ic_list}
\title{Reduce ic_list to include only pars with more than one value}
\usage{
reduce_ic_list(ic_list)
}
\arguments{
\item{ic_list}{}
}
\value{
list(ic_list, modified pars)
}
\description{
This can be used to clean the output, such that the condition description is not as long anymore.
}
\examples{

pars <- c(a=5,b=1)
ic_list <- c(a=1,b=1:2)
pars[names(reduce_ic_list(ic_list)[[2]])] <- reduce_ic_list(ic_list)[[2]]
ic_list <- reduce_ic_list(ic_list)[[1]]
}
