% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{ropt_gadget}
\alias{ropt_gadget}
\title{Shiny Gadget for quick exploration of Complexes and alpha}
\usage{
ropt_gadget(which_pars_opt, dMod.frame, hypothesis = 1)
}
\arguments{
\item{which_pars_opt}{All possible pars opt that you might want to include (you can kick them out later)}

\item{dMod.frame}{the dMod.frame with additional columns p_pert and pars}

\item{hypothesis}{indec to subset the dMod.frame}
}
\description{
Shiny Gadget for quick exploration of Complexes and alpha
}
