% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_preturbations.R
\name{run_different_perturbations}
\alias{run_different_perturbations}
\title{Run a set of dirreent perturbations over varying alphas and a parameter which can be varied}
\usage{
run_different_perturbations(dose_pars, pars0, alphas, perturbations, xs, p_log,
  p_pert, g, cores)
}
\arguments{
\item{pars0}{the "default" pars for the model}

\item{alphas}{vector of values for the free pars a to compare r_0 against r_a}

\item{perturbations}{list of sets of perturbations eg list(c(x1 = log(0.9), y1 = log(0.9)))}

\item{xs}{xs}

\item{p_log}{p_log}

\item{p_pert}{p_pert}

\item{g}{g}

\item{cores}{detectFreeCores()}

\item{u2sdose_pars}{vector of parameter values, they must be named}
}
\value{
a tibble
}
\description{
Run a set of dirreent perturbations over varying alphas and a parameter which can be varied
}
