% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{run_different_perturbations_noisy_with_replicates}
\alias{run_different_perturbations_noisy_with_replicates}
\title{Run the simulations and add noise to the data}
\usage{
run_different_perturbations_noisy_with_replicates(dose_pars, pars0,
  alpha_pars, alphas, perturbations, xs, p_log, p_pert, g, cores,
  alpha_par_settings, errorpars, N = 100, nreplicates = 3)
}
\arguments{
\item{dose_pars}{dosages eg  setNames(1:10, "x1")}

\item{pars0}{the "default" pars for the model}

\item{alpha_pars}{The default alpha parameters e.g. c(a_tox_Cxy = log(.Machine$double.eps))}

\item{alphas}{vector of values for the free pars a to compare r_0 against r_a}

\item{perturbations}{list of sets of perturbations eg list(c(x1 = log(0.9), y1 = log(0.9)))}

\item{xs}{xs}

\item{p_log}{p_log}

\item{p_pert}{p_pert}

\item{g}{g}

\item{cores}{detectFreeCores()}

\item{alpha_par_settings}{named list. Entries are vectors of the names of aplha_pars you want to use for the algorithm e.g. list(upstream = c("a_tox_Cxy", "a_toy_Cyz"))}

\item{errorpars}{list of vectors with absolute and relative errors: list(setting1 = c(sabs = 0, srel = 0.1))}

\item{N}{number of realizations for each setting}

\item{nreplicates}{number of replicates. normally distributed noise is assumed, replicates are averaged before starteing the algorithm}
}
\value{
tibble
}
\description{
Run the simulations and add noise to the data
}
\seealso{
[run_different_perturbations()]
}
