% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_ic_simulations.R
\name{run_simulations_with_noise}
\alias{run_simulations_with_noise}
\title{Run simulations with noise}
\usage{
run_simulations_with_noise(ic_list = NULL, n_samples = 100, sdlog = 0.05,
  cores = 3, pars_opt = pars_opt_0, obs_fun = g, pars = pars_0,
  pars_perturbed = pars_perturbed_0, fixed = NULL, alpha = 1)
}
\arguments{
\item{alpha}{}
}
\description{
Run simulations with noise
}

