% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonedfromdModandcOde.R
\name{runsteadyC}
\alias{runsteadyC}
\title{Interface to runsteady()}
\usage{
runsteadyC(y, times = c(0, Inf), func, parms, ...)
}
\arguments{
\item{y}{named vector of type numeric. Initial values for the integration}

\item{times}{vector of type numeric. Integration times}

\item{func}{return value from funC()}

\item{parms}{named vector of type numeric.}

\item{...}{further arguments going to \code{ode()}}
}
\value{
matrix with times and states
}
\description{
Interface to runsteady()
}
\details{
See deSolve-package for a full description of possible arguments
}

