% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{runtime}
\alias{runtime}
\title{The runtime of some code}
\usage{
runtime(code)
}
\arguments{
\item{code}{Any R code to be executed.}
}
\value{
The result of the code
}
\description{
The runtime of some code
}

