% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{scan_grid}
\alias{scan_grid}
\title{Title}
\usage{
scan_grid(which_par, dosages, alpha_scan_range = seq(-9, 9, by = 2),
  pars_opt = pars_opt_0, obs_fun = g, p_fun = (p_log * p_pert),
  pars = pars_0, grid_points_fine = alpha_scan_range,
  grid_points_raw = c(-9, 0, 2))
}
\arguments{
\item{grid_points_raw}{}
}
\description{
Title
}
