% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{steady_states_numerical}
\alias{steady_states_numerical}
\title{Simulate the System and take the last value of the simulation as the values for the steady states}
\usage{
steady_states_numerical(pars, time_max = 12000, predfun = (g * x))
}
\arguments{
\item{pars}{The parameters needed for the prediction function}

\item{time_max}{Initial guess at which time steady state is reached}

\item{predfun}{The prediction function. Right now, it should contain only one condition.}
}
\value{
The steady state concentrations that were obtained during the simulation.
}
\description{
The system is assumed to have reached steady state,
if the relative changes of the observables are less than 1e-3
}
\details{

}

