% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_states_dose_response.R
\name{steady_states_numerical_2}
\alias{steady_states_numerical_2}
\title{Simulate the System and take the last value of the simulation as the values for the steady states}
\usage{
steady_states_numerical_2(pars = pars_0, time_max = 12000, predfun = (g *
  x), by = 10)
}
\arguments{
\item{pars}{The parameters needed for the prediction function}

\item{time_max}{Initial guess at which time steady state is reached}

\item{predfun}{The prediction function. Right now, it should contain only one condition.}
}
\value{
The steady state concentrations that were obtained during the simulation.
 It is a condition-named list of steady_state concentrations with attributes counter_time_max =  c(counter, time_max)
 If count == 4, the predictions are also appended as attributes to look at the predictions, since in this case, the original time_max_guess has been already increased 1000 times, but steady state has not been reached.
}
\description{
The system is assumed to have reached steady state,
if the relative changes of the observables are less than 1e-3
}
\details{

}

