% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{svdinv}
\alias{svdinv}
\title{Inverse of matrix using singular value decomposition}
\usage{
svdinv(M, tol = .Machine$double.eps * 100)
}
\arguments{
\item{M}{Matrix}

\item{tol}{not used right now...}
}
\description{
Inverse of matrix using singular value decomposition
}
