% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_pars.R
\name{which_pars_opt}
\alias{which_pars_opt}
\title{Which pars_opt are associated with an element kept in the optimization procedure}
\usage{
which_pars_opt(pars_opt, r_kept, modules)
}
\arguments{
\item{modules}{}
}
\description{
Which pars_opt are associated with an element kept in the optimization procedure
}
