% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{dose_response_fun}
\alias{dose_response_fun}
\title{Compute dose-response curves}
\usage{
dose_response_fun(which_par, dosages, pars, predfun)
}
\arguments{
\item{which_par}{Character of length 1: Which par to perturb}

\item{dosages}{Numeric including the dosages eq(1,50,by = 1)}

\item{pars}{dynamical parameters}

\item{predfun}{prediction function, can be concatenated with parameter trafos and observation functions}
}
\description{
Compute the steady state dose-response wrt a parameter.
Return the dose response in a tidy data format.
}
