% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_states_dose_response.R
\name{dose_response_fun}
\alias{dose_response_fun}
\title{Compute dose-response curves}
\usage{
dose_response_fun(par_dose, dosages, pars, predfun = (g * x),
  depased_predfun = NULL)
}
\arguments{
\item{par_dose}{Character of length 1: Which par to perturb}

\item{dosages}{Numeric including the dosages eq(1,50,by = 1)}

\item{predfun}{}
}
\description{
Compute the steady state dose-response of a parameter.
Put the dose response out in a tidy data format.
}

