% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{local_response_matrix_eq10}
\alias{local_response_matrix_eq10}
\title{Calculate the local response matrix r from the global responses}
\usage{
local_response_matrix_eq10(R)
}
\arguments{
\item{R}{The matrix of global responses}
}
\value{
r The local response matrix, with the attribute which_pars_perturbed
}
\description{
Calculate the local response matrix r from the global responses
}
\details{

}

