% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{obj_alpha}
\alias{obj_alpha}
\title{Objective function to minimize certain elements of the local response matrix}
\usage{
obj_alpha(pars = pars_opt_0, fixed = NULL,
  perturbation_prediction = perturbation_prediction_0, r_kept = r_kept_0,
  obs_fun = g, p_fun = (p_log * p_pert), mypars = pars_0,
  method = "simple", ...)
}
\arguments{
\item{pars}{Free parameters a_ij to control the influence of the complexes}

\item{fixed}{Fixed parameters are parameters for which no derivative is calculated}

\item{perturbation_prediction}{Simulated steady state perturbation data}

\item{r_kept}{elements kept in the minimization (=elements that describe sequestration connections)}

\item{obs_fun}{Observation function which returns the communicating species}

\item{p_fun}{Parameter transformation which returns different conditions}

\item{mypars}{Parameters of the ODE model}

\item{method}{method for taking the derivative wrt pars_opt.}

\item{...}{not needed except for throwing away an unnecessary input when applying the new mstrust}
}
\description{
This function squares all elements that flip sign once the complexes are added to the observation.
This can then be used to minimize the values.
}
