% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{obj_alpha_no_derivs}
\alias{obj_alpha_no_derivs}
\title{Objective function which doesn't compute derivatives}
\usage{
obj_alpha_no_derivs(pars_opt, perturbation_prediction, r_kept, obs_fun, p_fun,
  mypars)
}
\arguments{
\item{mypars}{}
}
\description{
I used this function to compare different optimizers. trust() was as fast as optim, but more reliable.
All other methods like simulated annealing were not as reliable.
}
