% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{obj_alpha_no_derivs}
\alias{obj_alpha_no_derivs}
\title{Objective function which doesn't compute derivatives}
\usage{
obj_alpha_no_derivs(pars_opt, perturbation_prediction, r_kept, obs_fun,
  p_fun, mypars)
}
\arguments{
\item{perturbation_prediction}{Simulated steady state perturbation data}

\item{r_kept}{elements kept in the minimization (=elements that describe sequestration connections)}

\item{obs_fun}{Observation function which returns the communicating species}

\item{p_fun}{Parameter transformation which returns different conditions}

\item{mypars}{Parameters of the ODE model}
}
\description{
I used this function to compare different optimizers. trust() was as fast as optim, but more reliable.
All other methods like simulated annealing were not as reliable.
}
