% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{obj_frobenius}
\alias{obj_frobenius}
\title{Objective function to minimize certain elements of the local response matrix}
\usage{
obj_frobenius(pars_optimization = pars_opt_0, condition = 1,
  r_kept = list(rep(1, length(modules)^2)), pars = pars_0, predfun = (g *
  xs), which_pars_perturbed = which_pars_perturbed_0,
  r_fun = r_numDeriv_fun)
}
\arguments{
\item{which_pars_perturbed}{}
}
\value{
The parameter values for which the sign-changing elements are minimal.
}
\description{
This function squares all elements that flip sign once the complexes are added to the observation.
This can then be used to minimize the values.
}

