% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{obj_frobenius}
\alias{obj_frobenius}
\title{Objective function to minimize certain elements of the local response matrix}
\usage{
obj_frobenius(pars_optimization = pars_opt_0, pars = pars_0,
  time_max = 12000, predfun = (g * x),
  which_pars_perturbed = which_pars_perturbed_0, r_fun = r_numDeriv_fun)
}
\arguments{
\item{pars_optimization}{The parameters which are needed to optimize the matrix}
}
\value{
The parameter values for which the sign-changing elements are minimal.
}
\description{
This function squares all elements that flip sign once the complexes are added to the observation.
This can then be used to minimize the values.
}

