% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_generation.R
\name{p_pert_fun}
\alias{p_pert_fun}
\title{Convenient simulation of perturbation experiments}
\usage{
p_pert_fun(which_pars_perturbed = which_pars_perturbed_0,
  perturbation_strength = 1.1, pars = pars_0)
}
\arguments{
\item{which_pars_perturbed}{The names of the parameters to be perturbed.}

\item{perturbation_strength}{The fold-change of the parameter on linear scale.
Either one value to set all of them to the same strength or a numeric vector with the individual perturbation strengths.}

\item{pars}{All parameters of the model.}
}
\value{
A parameter trafo function of class parfn.
}
\description{
A parameter transformation function to conveniently generate conditions, in which on parameter was perturbed
}

