% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{r_alpha_fun}
\alias{r_alpha_fun}
\title{Parametric local response matrix}
\usage{
r_alpha_fun(pars_opt = pars_opt_0,
  perturbation_prediction = perturbation_prediction_0, obs_fun = g0,
  p_fun = (p_log * p_pert), pars = pars_0)
}
\arguments{
\item{pars_opt}{Free parameters a_ij to control the influence of the complexes}

\item{perturbation_prediction}{Simulated steady state perturbation data}

\item{obs_fun}{Observation function which returns the communicating species}

\item{p_fun}{Parameter transformation which returns different conditions}

\item{pars}{Parameters of the ODE model}
}
\description{
Parametric local response matrix
}
