% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{R_fun}
\alias{R_fun}
\title{Compute R from fractional changes of steady states}
\usage{
R_fun(pars_opt = pars_opt_0,
  perturbation_prediction = perturbation_prediction_0, obs_fun = g,
  p_fun = (p_log * p_pert), pars = pars_0)
}
\arguments{
\item{pars}{}
}
\description{
Compute R from fractional changes of steady states
}
