% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{R_fun_2}
\alias{R_fun_2}
\title{Compute R from fractional changes of steady states, but do the linear combination only after the derivatives have been taken}
\usage{
R_fun_2(pars_opt = pars_opt_0,
  perturbation_prediction = perturbation_prediction_0, obs_fun = g,
  p_fun = (p_pert), pars = pars_0)
}
\arguments{
\item{pars}{}
}
\description{
Compute R from fractional changes of steady states, but do the linear combination only after the derivatives have been taken
}
