% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{r_kept_fun}
\alias{r_kept_fun}
\title{Which elements to keep in the optimization procedure?}
\usage{
r_kept_fun(pars_opt = pars_opt_0,
  perturbation_prediction = perturbation_prediction_0, obs_fun = g,
  p_fun = p_pert, pars = pars_0)
}
\arguments{
\item{which_pars_perturbed}{}
}
\description{
Elements are kept if they are either small or if they change sign when switching on the complexes.
}
\details{
To do: have a look at what happens, if pars_opt = .5 or combinations of pars_opt (eg c(0,1), c(1,0) ...)
}

