% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{r_kept_fun2}
\alias{r_kept_fun2}
\title{Function to compare two local response matrices
in order to choose which elements to keep in the minimization process}
\usage{
r_kept_fun2(r_0, r_alpha)
}
\arguments{
\item{r_0}{calculated local response matrix with alpha = 0 (or logalpha -> -Inf)}

\item{r_alpha}{calculated alocal response  matrix with alpha = alpha}
}
\description{
Function to compare two local response matrices
in order to choose which elements to keep in the minimization process
}
