% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2Raf.R
\name{r_sens_fun}
\alias{r_sens_fun}
\title{Retrieve the local response matrix from the sensitivity equations}
\usage{
r_sens_fun(pars)
}
\arguments{
\item{pars}{The parameters for which r shall be computed}
}
\description{
Retrieve the local response matrix from the sensitivity equations
}

