% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_preturbations.R
\name{run_different_perturbations}
\alias{run_different_perturbations}
\title{Run a set of dirreent perturbations over varying alphas and a parameter which can be varied}
\usage{
run_different_perturbations(dose_pars, pars0, alpha_pars, alphas, perturbations,
  xs, p_log, p_pert, g, cores, alpha_par_settings)
}
\arguments{
\item{dose_pars}{dosages eg  setNames(1:10, "x1")}

\item{pars0}{the "default" pars for the model}

\item{alpha_pars}{The default alpha parameters e.g. c(a_tox_Cxy = log(.Machine$double.eps))}

\item{alphas}{vector of values for the free pars a to compare r_0 against r_a}

\item{perturbations}{list of sets of perturbations eg list(c(x1 = log(0.9), y1 = log(0.9)))}

\item{xs}{xs}

\item{p_log}{p_log}

\item{p_pert}{p_pert}

\item{g}{g}

\item{cores}{detectFreeCores()}

\item{alpha_par_settings}{named list. Entries are vectors of the names of aplha_pars you want to use for the algorithm e.g. list(upstream = c("a_tox_Cxy", "a_toy_Cyz"))}
}
\value{
a tibble
}
\description{
Run a set of dirreent perturbations over varying alphas and a parameter which can be varied
}
