% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_ic_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Run the optimization procedure at many conditions}
\usage{
run_simulations(ic_list = NULL, nfits = 10, pars_opt = pars_opt_0,
  obs_fun = g, pars = pars_0,
  which_pars_perturbed = which_pars_perturbed_0, fixed = NULL)
}
\arguments{
\item{p_ic_list}{}
}
\description{
Even though this function takes only one parameter as input, it depends on many objects to be defined.
}
\details{
xs(times, c(pars_raw, ic_raw))
(g*xs)(times, c(pars_raw, ic_raw, pars_opt))
pars_0, pars_opt_0

which_pars_perturbed_0, p_pert
}

