% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_states_dose_response.R
\name{steady_states_numerical}
\alias{steady_states_numerical}
\title{Steady states out of a simulation of a system}
\usage{
steady_states_numerical(pars = pars_0, predfun = (g * x),
  deparsed_predfun = NULL)
}
\arguments{
\item{pars}{The parameters for which the steady states are to be computed.}

\item{predfun}{An expression like (g*x*p), (g*x) or (x*p). Brackets can be omitted.}

\item{odemodel}{The object in which the odemodel is stored.}
}
\description{
This function gives access to the runsteady-function of deSolve within the formalism of dMod.
One can pass a prediction function like (g*x*p), (g*x) or (x*p).
The steady states of the ODE are computed by deSolve::runsteady and g and p are executed if they exist.
}

